Подтверждение удаления блока в редакторе

+++ public_html/admin/js/story-editor.js
@@ -793,6 +793,9 @@ var StoryEditor = (function() {
                     stretchToSlide();
                 },
                 'delete': function() {
+                    if (!confirm('Удалить блок?')) {
+                        return;
+                    }
                     deleteBlockAction();
                 },
                 'duplicate': function() {
